RemoteControl/requirements.txt
websocket-client
websockets
websocket_server